Logo concurrency and async threads util (#37)
* Added concurency to Logo finder

* Change amount of logos back to 3

Done some testing and it seems to be doing 3 regions in 2 seconds instead of 3 seconds. It takes 5 seconds instead of 7 seconds for 10 regions. So a small improvement.

* Add async_threads util ThreadWorker

ThreadWorker makes threads and handles running tasks concurrenltly.

* Add ProcessGroup to util async Threads

ProcessGroup handles keeping track of the tasks scheduled and returning them to you. Interacts with ThreadWorker.

* Update LogoFinder to use Util classes

Updated Util classes from async_threads. Worker is created (later on it can be passed on), using worker a new process_group is made and used for this part.

* answered questions

The standard amount of workers is fine for us. The Future we will use is concurrent.futures Future as it is thread safe and there is not much difference.

* Small improvements

* Renaming process to Future

* Added different return results of FutureGroup

Does not work perfectly yet, will continue tomorrow.

* Added worker to dst.py and now using 1 asyncio.run

The worker will allow running in the same threadpool. The usage of one asyncio.run is recommended. Using await instead on the places of asyncio.run Also changed the concurrency in two places.

* Added worker changes to Logo finder

* Add worker to logo finder

Avoiding using worker in arguments to logofinder, worker is initialized in logo finder itself seperately.

* Improve functions FutureGroup

The functions have been improved to use as_completed instead of a busy loop. Also the functions were tested and work.

* Improved asycn screenshots

Added cancel functionality, added the ability to use other functions besides =True for the any function.
Added 10 seconds of timeout delay instead of 5 seconds to taking screenshot.

* Addd preprocessor and screenshotter

Preprocessing with localdata allowing objects such as the screenshotter to be used in a seperate thread. Changed the arguments to work with it. All seems to work.

* Clean up async_thread

Cleaned up async_thread util classes. Shortened if tree by using expanding arguments.
Removed print statement, unneeded comment and renaming.

* added aiostream library and used it in logo search

* Documentation and small fixes

Based on the review comments of Teun, added documentation and fixed some of the issues. Other issues still need testing before commiting.

* Review improvements, documentation, using aiostream and default preprocessor

Last improvements from feedback PR review. Bit of documentation. Improved preprocessor using a default to not have if else statement. Lastly also added async_first and using stream for the screenshotter. This to fix resulturl and index having a correct return value available.